♻️ Refactor: Split aggregator creation by responsibility
