+++ frontend/src/components/dashboard/aggregator/aggregator.types.ts
@@ -1,0 +1,32 @@
+import { OptimizationResponse, AggregatorOption } from "@/api/aggregator";
+import { AggregatorOptimizeConfig, AggregatorConfig } from "@/api/aggregator";
+import {FederatedLearningData } from "@/api/aggregator";
+export interface CreationStatus {
+  step: "creating" | "selecting" | "deploying" | "completed" | "error";
+  message: string;
+  progress?: number;
+}
+
+export interface AggregatorSelectionModalProps {
+  results: OptimizationResponse;
+  onSelect: (option: AggregatorOption) => void;
+  onCancel: () => void;
+}
+
+export interface CreationStatusDisplayProps {
+  status: CreationStatus | null;
+}
+
+export interface ProgressStepsProps {
+  creationStatus: CreationStatus | null;
+  isLoading: boolean;
+}
+
+// Re-export types from API for convenience
+export type {
+  OptimizationResponse,
+  AggregatorOption,
+  FederatedLearningData,
+  AggregatorOptimizeConfig,
+  AggregatorConfig
+};
